+++ src/operations/fs/add.js
@@ -1,4 +1,4 @@
-import fs from 'fs/promises';
+import { writeFile } from 'fs/promises';
 import path from 'path';
 import { coloredLog } from '../../utils/getColoredLog.js';
 import { FILE_EXIST } from '../../constants/index.js';
@@ -6,7 +6,7 @@ import { FILE_EXIST } from '../../constants/index.js';
 export const add = async (fileName) => {
   try {
     const src = path.join(process.cwd(), fileName);
-    await fs.writeFile(src, '', { flag: 'wx+' });
+    await writeFile(src, '', { flag: 'wx+' });
     coloredLog(`File ${fileName} was successfully created`, 'green');
   } catch (error) {
     coloredLog(FILE_EXIST, 'red');
