+++ src/operations/fs/index.js
@@ -1,0 +1,6 @@
+export { cat } from './cat.js';
+export { add } from './add.js';
+export { rn } from './rn.js';
+export { cp } from './cp.js';
+export { mv } from './mv.js';
+export { rm } from './rm.js';
