+++ src/operations/navigation/index.js
@@ -1,0 +1,3 @@
+export { up } from './up.js';
+export { cd } from './cd.js';
+export { ls } from './ls.js';
