+++ src/operations/navigation/ls.js
@@ -1,4 +1,4 @@
-import fs from 'fs/promises';
+import { readdir, stat } from 'fs/promises';
 import path from 'path';
 import { coloredLog } from '../../utils/getColoredLog.js';
 import { SOMETHING_WRONG } from '../../constants/index.js';
@@ -6,13 +6,13 @@ import { SOMETHING_WRONG } from '../../constants/index.js';
 export const ls = async () => {
   try {
     const currentPath = process.cwd();
-    const items = await fs.readdir(currentPath);
+    const items = await readdir(currentPath);
 
     const promises = items.map(async (item) => {
       const fullPath = path.join(currentPath, item);
 
       try {
-        const stats = await fs.stat(fullPath);
+        const stats = await stat(fullPath);
         return {
           name: item,
           type: stats.isDirectory() ? 'directory' : 'file',
