--- src/utils/getHomeDir.js
@@ -1,3 +1,0 @@
-import { homedir } from 'os';
-
-export const getHomeDir = () => homedir();
