+++ src/utils/index.js
@@ -1,0 +1,6 @@
+export { getGreetingMsg } from './getGreetingMsg.js';
+export { getCurrentPathMsg } from './getCurrentPathMsg.js';
+export { getByeMsg } from './getByeMsg.js';
+export { parseArgs } from './parseArgs.js';
+export { getValueByArg } from './getValueByArg.js';
+export { coloredLog } from './getColoredLog.js';
